move linalg tensor to std
