+++ crates/cubecl-cuda/src/lib.rs
@@ -32,7 +32,6 @@ mod tests {
     cubecl_linalg::testgen_matmul_simple!([f16, bf16, f32]);
     cubecl_linalg::testgen_matmul_tiling2d!([f16, bf16, f32]);
     cubecl_linalg::testgen_conv2d_accelerated!([f16: f16, bf16: bf16, f32: tf32]);
-    cubecl_linalg::testgen_tensor_identity!([f16, bf16, f32, u32]);
     cubecl_reduce::testgen_reduce!([f16, bf16, f32, f64]);
     cubecl_random::testgen_random!();
     cubecl_reduce::testgen_shared_sum!([f16, bf16, f32, f64]);
