+++ crates/cubecl-linalg/src/lib.rs
@@ -6,6 +6,3 @@ pub mod matmul;
 
 /// Contains convolution using matmul components
 pub mod convolution;
-
-/// Contains basic tensor helpers.
-pub mod tensor;
