+++ crates/cubecl-linalg/src/matmul/kernels/tiling2d/launch.rs
@@ -2,13 +2,11 @@ use std::cmp::max;
 
 use cubecl_core::prelude::*;
 
-use crate::{
-    matmul::kernels::tiling2d::{
-        base::tiling2d_cube_kernel,
-        config::{CubeTiling2dConfig, tiling2d_cube_count, tiling2d_cube_dim},
-    },
-    tensor::{MatrixBatchLayout, TensorHandle, into_contiguous, matrix_batch_layout},
+use crate::matmul::kernels::tiling2d::{
+    base::tiling2d_cube_kernel,
+    config::{CubeTiling2dConfig, tiling2d_cube_count, tiling2d_cube_dim},
 };
+use cubecl_std::tensor::{MatrixBatchLayout, TensorHandle, into_contiguous, matrix_batch_layout};
 
 use super::config::Tiling2dConfig;
