--- crates/cubecl-linalg/src/tensor/mod.rs
@@ -1,13 +1,0 @@
-mod base;
-mod contiguous;
-pub mod identity;
-mod layout;
-
-pub use base::*;
-pub use contiguous::*;
-pub use identity::*;
-pub use layout::*;
-
-/// Tests for tensor kernels
-#[cfg(feature = "export_tests")]
-pub mod tests;
