--- crates/cubecl-std/src/tensor/base.rs
@@ -1,23 +1,0 @@
-/// Checks if the tensor associated with the given shape and strides is contiguous.
-pub fn is_contiguous(shape: &[usize], strides: &[usize]) -> bool {
-    if shape.is_empty() {
-        return true;
-    }
-
-    for (expected, &stride) in compact_strides(shape).into_iter().zip(strides) {
-        if expected != stride {
-            return false;
-        }
-    }
-
-    true
-}
-
-pub fn compact_strides(shape: &[usize]) -> Vec<usize> {
-    let rank = shape.len();
-    let mut strides = vec![1; rank];
-    for i in (0..rank - 1).rev() {
-        strides[i] = strides[i + 1] * shape[i + 1];
-    }
-    strides
-}
