+++ crates/cubecl-std/src/tensor/identity.rs
