+++ crates/cubecl-std/src/tensor/matrix_batch_layout.rs
