+++ crates/cubecl-std/src/tensor/mod.rs
@@ -1,4 +1,14 @@
-mod base;
+mod contiguous;
+mod handle;
+pub mod identity;
+mod matrix_batch_layout;
 
-pub use base::*;
+pub use contiguous::*;
+pub use handle::*;
+pub use identity::*;
+pub use matrix_batch_layout::*;
 pub mod r#virtual;
+
+// /// Tests for tensor kernels
+// #[cfg(feature = "export_tests")]
+// pub mod tests;
