+++ crates/cubecl-std/src/tests/tensor/identity.rs
