+++ crates/cubecl-std/src/tests/tensor/test_macros/identity.rs
@@ -9,29 +9,29 @@ macro_rules! testgen_tensor_identity {
     };
     ($numeric:ident) => {
             use super::*;
-            use cubecl_linalg::tensor::tests;
+            use $crate::tests::tensor::identity::test_identity;
             use cubecl_core::flex32;
 
             pub type NumericT = $numeric;
 
             #[test]
             pub fn test_tiny() {
-                cubecl_linalg::tensor::tests::identity::test_identity::<TestRuntime, NumericT>(&Default::default(), 3);
+                test_identity::<TestRuntime, NumericT>(&Default::default(), 3);
             }
 
             #[test]
             pub fn test_small() {
-                cubecl_linalg::tensor::tests::identity::test_identity::<TestRuntime, NumericT>(&Default::default(), 16);
+                test_identity::<TestRuntime, NumericT>(&Default::default(), 16);
             }
 
             #[test]
             pub fn test_normal() {
-                cubecl_linalg::tensor::tests::identity::test_identity::<TestRuntime, NumericT>(&Default::default(), 256)
+                test_identity::<TestRuntime, NumericT>(&Default::default(), 256)
             }
 
             #[test]
             pub fn test_large() {
-                cubecl_linalg::tensor::tests::identity::test_identity::<TestRuntime, NumericT>(&Default::default(), 4096)
+                test_identity::<TestRuntime, NumericT>(&Default::default(), 4096)
             }
     };
     ([$($numeric:ident),*]) => {
