+++ crates/cubecl-std/src/tests/tensor/test_macros/mod.rs
